backend/_DESCARTADOS/AIProcessingService/workers/requirements.txt
# =============================================================================
# Professional Vehicle Segmentation Pipeline - Requirements
# =============================================================================
# 7-Stage Pipeline: Detection → Shadow → SAM2 → CarParts → SegFormer → ISNet → Enhancement
# =============================================================================

# -----------------------------------------------------------------------------
# Core Dependencies
# -----------------------------------------------------------------------------
numpy>=1.24.0
Pillow>=10.0.0
scipy>=1.11.0
httpx>=0.25.0
aio-pika>=9.3.0
boto3>=1.34.0
requests>=2.31.0
tqdm>=4.66.0

# -----------------------------------------------------------------------------
# OpenCV for image processing and mask refinement
# -----------------------------------------------------------------------------
opencv-python>=4.8.0

# -----------------------------------------------------------------------------
# ML Frameworks (PyTorch ecosystem)
# -----------------------------------------------------------------------------
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# -----------------------------------------------------------------------------
# HuggingFace Transformers (GroundingDINO, SegFormer, CarParts)
# -----------------------------------------------------------------------------
transformers>=4.36.0
huggingface_hub>=0.20.0

# -----------------------------------------------------------------------------
# Stage 1: YOLOv8x-seg (Ultralytics)
# -----------------------------------------------------------------------------
ultralytics>=8.1.0

# -----------------------------------------------------------------------------
# Stage 1: GroundingDINO (zero-shot detection)
# Loaded via transformers (IDEA-Research/grounding-dino-base)
# -----------------------------------------------------------------------------

# -----------------------------------------------------------------------------
# Stage 2: ShadowFormer (ONNX Runtime)
# -----------------------------------------------------------------------------
onnxruntime>=1.16.0
onnxruntime-gpu>=1.16.0  # Optional: for GPU acceleration

# -----------------------------------------------------------------------------
# Stage 3: Meta SAM2 (Segment Anything 2)
# -----------------------------------------------------------------------------
git+https://github.com/facebookresearch/segment-anything-2.git

# -----------------------------------------------------------------------------
# Stage 4: CarParts-Seg (HuggingFace model)
# Model: Armandoliv/cars-parts-segmentation-unet-resnet18
# Loaded via transformers
# -----------------------------------------------------------------------------

# -----------------------------------------------------------------------------
# Stage 5: SegFormer (NVIDIA/HuggingFace)
# Model: nvidia/segformer-b5-finetuned-ade-640-640
# Loaded via transformers
# -----------------------------------------------------------------------------

# -----------------------------------------------------------------------------
# Stage 6: ISNet/DIS (custom implementation in isnet.py)
# Weights: isnet-general-use.pth from xuebinqin/DIS
# -----------------------------------------------------------------------------

# -----------------------------------------------------------------------------
# Stage 7: Enhancement (RealESRGAN + CodeFormer)
# -----------------------------------------------------------------------------
basicsr>=1.4.2
realesrgan>=0.3.0
facexlib>=0.3.0

# -----------------------------------------------------------------------------
# OpenAI CLIP (for auxiliary tasks)
# -----------------------------------------------------------------------------
git+https://github.com/openai/CLIP.git

# -----------------------------------------------------------------------------
# Utilities
# -----------------------------------------------------------------------------
python-dotenv>=1.0.0
pydantic>=2.5.0
gdown>=5.0.0  # For Google Drive downloads

# -----------------------------------------------------------------------------
# Optional: Monitoring and Logging
# -----------------------------------------------------------------------------
# prometheus_client>=0.19.0
# loguru>=0.7.0
